+++ src/components/Header/utils/encoreCatalogLogOutTimer.ts
@@ -1,6 +1,7 @@
 import Cookies from "js-cookie";
 
-export const encoreLogOutURL =
+// TODO
+export const catalogLogOutURL =
   "https://browse.nypl.org/iii/encore/logoutFilterRedirect?suite=def";
 // Expiration time is 30 minutes set in milliseconds.
 export const patLoggedInCookieExpiredTime = 1800000;
@@ -51,6 +52,8 @@ function EncoreCatalogLogOutTimer(
       "catalog.nypl.org",
       // the domain of Test Research Catalog's pages
       "nypl-sierra-test.nypl.org",
+      // new domain
+      "borrow.nypl.org",
     ];
     const isOnValidDomain = domainsForExtendingLogIn.some(
       (domain) => domain === currentLocationHost
@@ -74,9 +77,12 @@ function EncoreCatalogLogOutTimer(
         const lastVisitedValidDomainTime = Cookies.get(
           "VALID_DOMAIN_LAST_VISITED"
         );
-        const timeUntilLogOut = lastVisitedValidDomainTime["VALID_DOMAIN_LAST_VISITED"]
+        const timeUntilLogOut = lastVisitedValidDomainTime[
+          "VALID_DOMAIN_LAST_VISITED"
+        ]
           ? patLoggedInCookieExpiredTime -
-            (currentTime - lastVisitedValidDomainTime["VALID_DOMAIN_LAST_VISITED"])
+            (currentTime -
+              lastVisitedValidDomainTime["VALID_DOMAIN_LAST_VISITED"])
           : undefined;
 
         this.logOutFromEncoreAndCatalogIn(timeUntilLogOut);
@@ -122,7 +128,7 @@ function EncoreCatalogLogOutTimer(
     const iframeExistingTime = 10000;
 
     // Set the `iframe`'s src to the endpoint URL for logging out from Encore.
-    logoutIframe.setAttribute("src", encoreLogOutURL);
+    logoutIframe.setAttribute("src", catalogLogOutURL);
     // Assign an ID for CSS usage.
     logoutIframe.setAttribute("id", "logoutIframe");
     logoutIframe.style.display = "none";