+++ src/components/Header/utils/headerUtils.ts
@@ -19,6 +19,7 @@ const authServerBase = {
   development: "https://dev-login.nypl.org/auth",
 };
 const baseLoginLinks = {
+  // TODO:
   catalog: {
     production: "https://browse.nypl.org/iii/encore/myaccount",
     development: "https://nypl-encore-test.nypl.org/iii/encore/myaccount",
@@ -34,7 +35,7 @@ export const tokenRefreshLink = (isProduction = true) => {
   return `${authServerBase[type]}/refresh`;
 };
 // Return the proper links when logged in or not. These are for the NYPL
-// "Log in" or "Go to" links for Encore and the Research Catalog. This is
+// "Log in" or "Go to" links for the Catalog and the Research Catalog. This is
 // based on the environment and logged in status.
 export const getLoginLinks = (patronName = "", isProduction = true) => {
   const type = isProduction ? "production" : "development";
@@ -113,23 +114,23 @@ export const siteNavLinks = [
 const generateQueriesForTracking = () => {
   // the time stamp here is for the purpose of telling when this search query is made.
   const currentTimeStamp = new Date().getTime();
-  return `?searched_from=header_search&timestamp=${currentTimeStamp}`;
+  return `searched_from=header_search&timestamp=${currentTimeStamp}`;
 };
 
 /**
- * Returns the final URL for the NYPL Encore search.
+ * Returns the final URL for the NYPL Catalog search.
  */
-export const getEncoreCatalogURL = (searchValue) => {
+export const getCatalogURL = (searchValue) => {
   const encodedSearchInput = encodeURIComponent(searchValue);
-  const rootUrl = "https://browse.nypl.org/iii/encore/search/";
-  let finalEncoreUrl;
+  const rootUrl = "https://borrow.nypl.org/search";
+  let finalUrl;
 
   if (encodedSearchInput) {
-    finalEncoreUrl =
-      `${rootUrl}C__S${encodedSearchInput}__Orightresult__U` +
+    finalUrl =
+      `${rootUrl}?query=${encodedSearchInput}&searchType=everything&pageSize=10&` +
       generateQueriesForTracking() +
       `&lang=eng`;
-    return finalEncoreUrl;
+    return finalUrl;
   }
   return null;
 };
@@ -158,6 +159,7 @@ export const getNYPLSearchURL = (searchString) => {
     return (
       catalogUrl +
       encodeURIComponent(searchString) +
+      "?" +
       generateQueriesForTracking()
     );
   }